Fix narrowing conversion of connection ID (#12264)
ProxySession stores the connection ID as an int64_t, but in the HttpSM
and HttpUserAgent, it was stored and transferred as an int type
(typically 32 bits). This resulted in unnecessary narrowing. This
updates the state machine to use an 64 bit type to alleviate the
narrowing.

(cherry picked from commit 6a6db7b7eaa5642a27e39da6de47a0c05a147366)

+++ include/proxy/http/HttpSM.h
@@ -614,10 +614,10 @@ public:
   bool set_server_session_private(bool private_session);
   bool is_dying() const;
 
-  int client_connection_id() const;
-  int client_transaction_id() const;
-  int client_transaction_priority_weight() const;
-  int client_transaction_priority_dependence() const;
+  int64_t client_connection_id() const;
+  int     client_transaction_id() const;
+  int     client_transaction_priority_weight() const;
+  int     client_transaction_priority_dependence() const;
 
   ink_hrtime get_server_inactivity_timeout();
   ink_hrtime get_server_active_timeout();
@@ -672,7 +672,7 @@ HttpSM::is_dying() const
   return terminate_sm;
 }
 
-inline int
+inline int64_t
 HttpSM::client_connection_id() const
 {
   return _ua.get_client_connection_id();
